requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
azure-ai-generative==1.0.0b11
azure-ai-ml==1.27.1
azure-ai-resources==1.0.0b9
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-speech==1.44.0
azure-common==1.1.28
azure-core==1.34.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.23.0
azure-mgmt-core==1.5.0
azure-mgmt-resource==22.0.0
azure-monitor-opentelemetry==1.6.9
azure-monitor-opentelemetry-exporter==1.0.0b36
azure-storage-blob==12.25.1
azure-storage-file-datalake==12.20.0
azure-storage-file-share==12.21.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
credentials==1.1
cryptography==45.0.2
databricks-sdk==0.54.0
Deprecated==1.2.18
distro==1.9.0
dotenv==0.9.9
etils==1.12.2
fastapi==0.115.12
filelock==3.18.0
fixedint==0.1.6
flax==0.10.6
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
yt-dlp==2025.5.22
zipp==3.21.0
openai==1.81.0
uvicorn==0.34.2
python-dotenv==1.1.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
scipy==1.15.3
setuptools==80.8.0
soundfile==0.13.1
starlette==0.46.2
httpcore==1.0.9
httpx==0.28.1
typing_extensions==4.13.2
python-multipart==0.0.20

